separate models from routes and temporary test improvements
